# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Garage API",
        "description": "Documentation de l'API Gestion Garage",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/mecaniciens": {
            "get": {
                "tags": [
                    "Mecaniciens"
                ],
                "summary": "Lister les mecaniciens",
                "operationId": "4e039f38de2deca332d02d0879cc93b1",
                "responses": {
                    "200": {
                        "description": "Liste des mecaniciens recuperee avec succes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mecaniciens"
                ],
                "summary": "Creer un mecanicien",
                "operationId": "550143570673a7158cbd29abbf29458d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "specialite"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Camara"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Ibrahima"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "624000000"
                                    },
                                    "specialite": {
                                        "type": "string",
                                        "example": "Moteur diesel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mecanicien cree avec succes"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/mecaniciens/{id}": {
            "get": {
                "tags": [
                    "Mecaniciens"
                ],
                "summary": "Afficher un mecanicien",
                "operationId": "de9d0cdbab2b3e46b5ce345217882137",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du mecanicien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mecanicien recupere avec succes"
                    },
                    "404": {
                        "description": "Mecanicien introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mecaniciens"
                ],
                "summary": "Modifier un mecanicien",
                "operationId": "9466f4ad87fda62990b6fd2c8a5d3f36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du mecanicien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "specialite"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Bah"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Moussa"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "628111111"
                                    },
                                    "specialite": {
                                        "type": "string",
                                        "example": "Electricite auto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mecanicien modifie avec succes"
                    },
                    "404": {
                        "description": "Mecanicien introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mecaniciens"
                ],
                "summary": "Supprimer un mecanicien",
                "operationId": "d994d707cb2d8c0339ec2a82f2b90ea7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du mecanicien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mecanicien supprime avec succes"
                    },
                    "404": {
                        "description": "Mecanicien introuvable"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Mecaniciens"
                ],
                "summary": "Modifier partiellement un mecanicien",
                "operationId": "a8bf1d84bb00734fd2915efe42fb90f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du mecanicien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Bah"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Moussa"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "628111111"
                                    },
                                    "specialite": {
                                        "type": "string",
                                        "example": "Electricite auto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mecanicien modifie avec succes"
                    },
                    "404": {
                        "description": "Mecanicien introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
...
}